Tests: Explicit cast to int in RandomImageGenerator test (php8 warnings)
Otherwise this causes E_DEPRECATED warnings to show up in the
AbuseFilter tests on php 8.1.

(cherry picked from commit 8cee8432b21190cb178285cd3ca4518fe49f0c24)

+++ tests/phpunit/includes/api/RandomImageGenerator.php
@@ -173,7 +173,7 @@ class RandomImageGenerator {
 
 		$draws = [];
 		for ( $i = 0; $i <= $this->shapesToDraw; $i++ ) {
-			$radius = mt_rand( 0, $diagonalLength / 4 );
+			$radius = mt_rand( 0, (int)( $diagonalLength / 4 ) );
 			if ( $radius == 0 ) {
 				continue;
 			}
